server/requirements.txt
Flask==3.0.3
werkzeug==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
PyMYSQL==1.1.1
